requirements.txt
pandas
numpy
matplotlib
seaborn
sqlalchemy
python-dotenv